feat/#56: Contract 분리
